App_WineQualityClass/requirements.txt
Flask
numpy  
scipy 
scikit-learn==0.23.1 
lxml
pandas==0.23.4
fasttext
tensorflow==1.4.0
keras==2.2.4
Pillow
fasttext
scipy==1.2.1
nltk==3.5
fasttext==0.9.2